Master should store the list of completed framework ids for lifecycle.
It should be separately from that for webUI and endpoints.
Currently the master stores the history of completed frameworks in
a map with the full historical data of the framework, it is
prohibitively expensive to keep a long history; In order to reject
frameworks from reregistering if they have previously marked as
completed, we only need to persist the framework IDs and are able
to keep long history.

Review: https://reviews.apache.org/r/71080